Issue-171 - Fixed problems found in the android studio inspect code

+++ client/src/main/java/com/github/goldy1992/mp3player/client/ui/components/equalizer/smoothline/SmoothLineEqualizer.kt
@@ -4,11 +4,18 @@ import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.runtime.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.CornerRadius
 import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.*
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.Path
+import androidx.compose.ui.graphics.StrokeCap
+import androidx.compose.ui.graphics.asAndroidPath
+import androidx.compose.ui.graphics.asComposePath
 import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.unit.dp
@@ -120,7 +127,7 @@ fun SmoothLineEqualizer(modifier: Modifier = Modifier,
                     lineColor,
                     Color.Transparent,
                 ),
-                endY = lineHeight.toFloat()
+                endY = lineHeight
             ),
         )
         drawPath(
